+++ src/OpenTelemetry/Trace/TracerProviderSdk.cs
@@ -34,15 +34,10 @@ namespace OpenTelemetry.Trace
         private readonly ActivitySourceAdapter adapter;
         private BaseProcessor<Activity> processor;
 
-        static TracerProviderSdk()
-        {
-            Activity.DefaultIdFormat = ActivityIdFormat.W3C;
-            Activity.ForceDefaultIdFormat = true;
-        }
-
         internal TracerProviderSdk(
             Resource resource,
             IEnumerable<string> sources,
+            IEnumerable<TracerProviderBuilder.DiagnosticSourceInstrumentationFactory> diagnosticSourceInstrumentationFactories,
             IEnumerable<TracerProviderBuilder.InstrumentationFactory> instrumentationFactories,
             Sampler sampler,
             List<BaseProcessor<Activity>> processors)
@@ -55,12 +50,20 @@ namespace OpenTelemetry.Trace
                 this.AddProcessor(processor);
             }
 
-            if (instrumentationFactories.Any())
+            if (diagnosticSourceInstrumentationFactories.Any())
             {
                 this.adapter = new ActivitySourceAdapter(sampler, this.processor, resource);
-                foreach (var instrumentationFactory in instrumentationFactories)
+                foreach (var instrumentationFactory in diagnosticSourceInstrumentationFactories)
                 {
                     this.instrumentations.Add(instrumentationFactory.Factory(this.adapter));
+                }
+            }
+
+            if (instrumentationFactories.Any())
+            {
+                foreach (var instrumentationFactory in instrumentationFactories)
+                {
+                    this.instrumentations.Add(instrumentationFactory.Factory());
                 }
             }