+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 0.8.0-beta.1
+
+Released 2020-Nov-5
+
 * ZipkinExporter will now respect global Resource set via
   `TracerProviderBuilder.SetResource`.
   ([#1385](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1385))