+++ src/OpenTelemetry.Extensions.Hosting/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 0.8.0-beta.1
+
+Released 2020-Nov-5
+
+* Removed AddOpenTelemetryTracing method which takes Func returning
+  TracerProvider.
+
 ## 0.7.0-beta.1
 
 Released 2020-Oct-16