+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 0.8.0-beta.1
+
+Released 2020-Nov-5
+
 * Record `Exception` in AspNetCore instrumentation based on `RecordException` in
   `AspNetCoreInstrumentationOptions`
   ([#1408](https://github.com/open-telemetry/opentelemetry-dotnet/issues/1408))
@@ -14,6 +18,11 @@
   to CompositeTextMapPropagator. IPropagator is renamed to TextMapPropagator
   and changed from interface to abstract class.
   ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1427))
+* Propagators.DefaultTextMapPropagator will be used as the default Propagator
+  ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1428))
+* Removed Propagator from Instrumentation Options. Instrumentation now always
+  respect the Propagator.DefaultTextMapPropagator.
+  ([#1448](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1448))
 
 ## 0.7.0-beta.1